backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
faker
python-dotenv
openai
